Fix versions for all python dependencies
This is a reminder that _all_ the dependencies should be recorded when
creating the Docker image, even the transitive ones. In this case, I was
transitively getting a different version of numpy than two years ago,
and this broke the code, even if I was not relying directly on anything
specific from numpy.

Hence, friendly reminder:

    pip freeze > requirements.txt